+++ source/bin/secure_media_stream.ts
@@ -7,8 +7,8 @@ import { AutoSessionRevocationStack, AutoSessionRevocationStackProps } from "../
 import { IConfiguration } from "../helpers/validators/configuration";
 
 const solutionId = 'SO0195';
-const solutionDisplayName = 'Secure Media Delivery at the Edge';
-const solutionVersion = '1.0.1';
+const solutionDisplayName = 'Secure Media Delivery at the Edge on AWS';
+const solutionVersion = '1.1.0';
 const description = `(${solutionId}) - ${solutionDisplayName}. Version ${solutionVersion}`;
 
 const app = new App();